marketwindow 3.4 look same like 2.7

+++ farmlib/marketwindow.py
@@ -59,7 +59,7 @@ class MarketWindow(Container):
         """
         print('generate Marketwindow')
         posx, posy = [0, 0]
-        columns = (self.width / 64) - 1
+        columns = int(self.width / 64) - 1
         # rows = (self.height / 32) - 1
         # Background
         bgimage = self.imgloader["marketbg"]
